pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Conversational RAG chatbot over PDF with Langfuse observability"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pdfplumber>=0.11.0",
    "chromadb>=0.5.0",
    "openai>=1.40.0",
    "langfuse>=2.50.0",
    "streamlit>=1.38.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "Pillow>=10.0.0",
    "onnxruntime<1.20",
]

[project.scripts]
ingest = "rag_chatbot.ingest:main"
evaluate = "rag_chatbot.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_chatbot"]
